Refactor: firebase config와 utils 함수 선언 분리
